changed qsub to run only if file does not exist

+++ deep/qsub.py
@@ -1,4 +1,5 @@
 import os,sys,re,time
+import os.path
 
 bash_script = '''#!/bin/bash
 #PBS -r n
@@ -17,10 +18,12 @@ python main.py {} {} {}
 for run_number in range(200):
 	for k in [0.05,0.75,0.1,0.15,0.2,0.25,0.35,0.5,1.0,2.0]:
 		for variance in [0.1,0.05]:
-			outfile="EM_{}_{}_{}.pbs".format(str(run_number),str(k),str(variance))
-			output=open(outfile, 'w')
-			print >>output, (bash_script.format(str(run_number),str(k),str(variance)))
-			output.close()
-			cmd="qsub -l short %s" % outfile
-			os.system(cmd)
-			time.sleep(.25)
+			fname="w_loss-"+str(run_number)+"-"+str(k)+"-"+str(variance)+".txt"
+			if os.path.isfile(fname)==False:
+				outfile="EM_{}_{}_{}.pbs".format(str(run_number),str(k),str(variance))
+				output=open(outfile, 'w')
+				print >>output, (bash_script.format(str(run_number),str(k),str(variance)))
+				output.close()
+				cmd="qsub -l short %s" % outfile
+				os.system(cmd)
+				time.sleep(.5)
